Removed Unused Splitter and Created New Splitter Module
